refactor: add shared/logic and move logic codes
