+++ apps/user/src/routes/register/+page.server.ts
@@ -1,11 +1,11 @@
 import type { Actions, PageServerLoad } from './$types';
 import { i18n } from '$lib/i18n';
 import * as auth from '$lib/server/auth';
-import { checkUserEmailAvailability, verifyEmailInput } from '$lib/server/email';
-import * as ev from '$lib/server/email-verification';
-import { hashPassword, verifyPasswordInput, verifyPasswordStrength } from '$lib/server/password';
 import { db } from '@shared/db';
 import * as table from '@shared/db/schema';
+import { checkUserEmailAvailability, verifyEmailInput } from '@shared/logic/email';
+import * as ev from '@shared/logic/email-verification';
+import { hashPassword, verifyPasswordInput, verifyPasswordStrength } from '@shared/logic/password';
 import { fail, redirect } from '@sveltejs/kit';
 
 export const load: PageServerLoad = async (event) => {
