+++ shared/logic/email.ts
