+++ shared/logic/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@shared/logic",
+  "type": "module",
+  "version": "0.0.1",
+  "main": "index.js",
+  "dependencies": {
+    "@node-rs/argon2": "^2.0.2",
+    "@oslojs/crypto": "^1.0.1",
+    "@oslojs/encoding": "^1.1.0",
+    "drizzle-orm": "^0.37.0",
+    "std-env": "^3.8.0"
+  },
+  "devDependencies": {
+    "@sveltejs/kit": "^2.9.0"
+  }
+}
