+++ shared/logic/utils.ts
